savegame: Fix crash with illegal action lists
During the save or load of a game, if an order's action list is unapproved,
or malformed, the internal memory for the action list is deleted and
set to NULL.

But the deleted memory is later accessed as part of a sanity test.

Added defensive code to guard against access of the NULL pointer.

See RM #1646